Issue #29 Set correct htpassword file in Redhat
On Redhat-like systems, the default location for the Nagios htpasswd
file is /etc/nagios/passwd, which was not being correctly set by
puppet-nagios.  Updated params.pp to use that new location.  Also
updated params.pp to use the osfamily fact for that test instead of
using operatingsystem.

Added appropriate tests.